refactor/updated the code according to comments
