+++ src/components/shared/PasswordInput/index.tsx
@@ -8,6 +8,7 @@ import StyledPasswordInput from './styledPasswordInput';
 const PasswordInput = forwardRef<HTMLInputElement, StyledInputProps>(
   ({ placeholder, width, padding, name, stylevariant, ...props }, ref) => {
     const [showPassword, setShowPassword] = useState(false);
+
     const handleClickShowPassword = () => setShowPassword(!showPassword);
     const handleMouseDownPassword = (
       event: React.MouseEvent<HTMLButtonElement>
